+++ metadata.rb
@@ -1,10 +1,10 @@
+name             'debian'
 maintainer       'Reaktor Innovations Oy'
 maintainer_email 'cookbooks@reaktor.fi'
 license          'Apache v2.0'
-name             'debian'
 description      'Sets up Debian repositories'
-long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.2.1'
+version          '1.3.0'
 
 supports         'debian'
+
 depends          'apt', '>= 1.6.0'